requirements.txt
requests
arxiv
sentence-transformers
faiss-cpu
numpy
tqdm
pydantic
scikit-learn
scholarly